Code refactoring
An analysis of the JS code was made, according to the ESLint style guide.
Separate style files were created for each of the components that were established in the app.
New icons were implemented for the app.
Added propotype for README.md
