+++ recetario/App.js
@@ -17,7 +17,7 @@ const App = () => {
         headerShown: false
       }}
     >
-    <Stack.Screen name="Home" component={Home}  />
+    <Stack.Screen name="Home" component={Home} />
     <Stack.Screen name="Menu" component={Menu} />
 
     </Stack.Navigator>
