+++ recetario/components/Ingredient.js
@@ -1,29 +1,14 @@
 import React from 'react'
-import { Text, View, StyleSheet } from 'react-native'
+import { Text, View } from 'react-native'
+import { ingredientsStyles } from '../styles/ingredientsStyles'
 
 const Ingredient = ({ item }) => {
-  return(
-  <View style={styles.contentIngredient}>
-    <Text style={styles.ingredient}>{item.name}</Text>
-    <Text style={styles.ingredient}>{item.size}</Text>
+  return (
+  <View style={ingredientsStyles.contentIngredient}>
+    <Text style={ingredientsStyles.ingredient}>{item.name}</Text>
+    <Text style={ingredientsStyles.ingredient}>{item.size}</Text>
   </View>
   )
 }
-
-const styles = StyleSheet.create({
-  contentIngredient: {
-    height: 30,
-    alignItems: 'flex-start',
-    flexDirection: "row",
-    justifyContent: "space-between",
-    borderBottomColor: '#3f3f3f',
-    borderBottomWidth: 1,
-    marginHorizontal: 15,
-    marginVertical: 5
-  },
-  ingredient: {
-    color: '#efedec'
-  }
-})
 
 export default Ingredient
