+++ recetario/styles/appStyles.js
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native'
+
+const appStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'column',
+    backgroundColor: '#282828',
+    paddingHorizontal: '2.5%'
+  },
+  text: {
+    color: '#d31c71',
+    fontSize: 24,
+    marginTop: 10,
+    marginBottom: 10
+  },
+  search: {
+    marginTop: 30
+  }
+})
+
+export { appStyles }
