+++ recetario/styles/horizontalStyles.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native'
+
+const horizontalStyles = StyleSheet.create({
+  horizontal: {
+    height: 250,
+    width: '100%',
+    flexDirection: 'row'
+  },
+  text: {
+    color: '#d31c71',
+    fontSize: 24
+  }
+})
+
+export { horizontalStyles }
