+++ recetario/styles/itemStyles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native'
+
+const itemStyles = StyleSheet.create({
+  item: {
+    height: 225,
+    width: 180
+
+  },
+  img: {
+    height: 150,
+    width: 150,
+    borderRadius: 5
+  },
+  text: {
+    color: 'white'
+  }
+})
+
+export { itemStyles }
